+++ Runtime/ColorEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Channels
+{
+    [CreateAssetMenu(menuName = "Channels/Color Event Channel")]
+    public class ColorEventChannel : EventChannel<Color> { }
+}