+++ Runtime/DoubleEventChannel.cs
@@ -5,7 +5,7 @@ namespace UnderLogic.Channels
     [CreateAssetMenu(menuName = "Channels/Double Event Channel")]
     public class DoubleEventChannel : EventChannel<double>
     {
-        public void RaiseEvent(int value) => RaiseEvent((double)value);
-        public void RaiseEvent(float value) => RaiseEvent((double)value);
+        public void RaiseEvent(int value) => base.RaiseEvent(value);
+        public void RaiseEvent(float value) => base.RaiseEvent(value);
     }
 }