+++ Runtime/FloatEventChannel.cs
@@ -5,6 +5,6 @@ namespace UnderLogic.Channels
     [CreateAssetMenu(menuName = "Channels/Float Event Channel")]
     public class FloatEventChannel : EventChannel<float>
     {
-        public void RaiseEvent(int value) => RaiseEvent((float)value);
+        public void RaiseEvent(int value) => base.RaiseEvent(value);
     }
 }