+++ Runtime/Listeners/ColorEventListener.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Channels.Listeners
+{
+    public class ColorEventListener : EventListener<Color, ColorEventChannel> { }
+}