+++ Runtime/Listeners/Vector2EventListener.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace UnderLogic.Channels.Listeners
+{
+    public class Vector2EventListener : EventListener<Vector2, Vector2EventChannel>
+    {
+        public void RaiseEvent(Vector2Int value) => base.RaiseEvent(new Vector2(value.x, value.y));
+    }
+}