+++ Runtime/Listeners/Vector2IntEventListener.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Channels.Listeners
+{
+    public class Vector2IntEventListener : EventListener<Vector2Int, Vector2IntEventChannel> { }
+}