+++ Runtime/Vector2IntEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Channels
+{
+    [CreateAssetMenu(menuName = "Channels/Vector2Int Event Channel")]
+    public class Vector2IntEventChannel : EventChannel<Vector2Int> { }
+}