+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2022-12-04
+
+### Added
+
+- `ColorEventChannel` variant
+- `ColorEventListener` component
+- `Vector2IntEventChannel` variant
+- `Vector2IntEventListener` component
+- `Vector2EventChannel` variant
+- `Vector2EventListener` component
+- `Vector3EventChannel` variant
+- `Vector3EventListener` component
+- `Vector4EventChannel` variant
+- `Vector4EventListener` component
+
 ## [0.3.1] - 2022-12-04
 
 ### Added